requirements.txt
# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain y LLMs
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
langchain-openai==0.0.2
langchain-core==0.1.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Utilidades
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Logging y monitoreo
structlog==23.2.0
